Allow downgrading guest MTU settings
Most standard networks will have a 1500 MTU sized
provider network.  Because there is a 24 byte overhead
with using any kind of encapsulation, most users will
want to downgrade the MTU set within their guests or
raise their internal network MTU above 1500.  The former
is much easier to do and has been added as an option to
dnsmasq that is passed to the client.

Change-Id: I199ec6e46c114d0277230eded002ad99786c061c