Allow building OVS from souce
The lastest version of OVS in Ubuntu 12.04 is
quite old (1.4). 1.9 is the new LTS supported
version and 1.11 is the newest head version.

Version 1.10/1.11 has many new features and
performance improvements, so there is a benefit
in being able to install it over the default 1.4
packages.

By default, nothing changes in this patch.
You must explicitly set
['openstack']['network']['openvswitch']['use_source_version']
to true in order to enable it.

Change-Id: I3779b6b73f3e2c0a56d3bed6bf3e3a4184649c24