Add check for the external interface
Because ovs-vsctl add-port simply updates the OVS db, it will
dutifully add it to the configuation even if the external
interface doesn't exist on the system.

The end result of this is that traffic will silently die as it
attempts to go to the wire.

This check verifies the external interface is present on the system
and doesn't attempt to add it to the external bridge unless it exists.

Change-Id: Ic93e21e34bc14fac06e764e06f335630c3c6eac4