Add takeover delay to quantum-ha-tool
Add a takeover delay to the quantum-ha-tool to ensure we do
not try to aggressively takeover.  We want to ensure an agent
is offline for some period of time (TAKEOVER_DELAY seconds)
before we begin a migration.  We also randomize it a bit in the
event that this script is being run in parallel from several hosts.

Change-Id: Ied9a75791b27af310f81f4cec327b3699fbae7ed