fixed the error when restarting quantum-server
When setup a network node which is not include
the quantum-server daemon by default, there will
be an error in the recipes ./openstack-network/recipes/
openvswitch.rb when retarting quantum-server service.
This error doesn't affect the whole procedure. My fixes
will check whether the quantum-server in the role-list.
It will not output the error again.

Change-Id: Iad276aaafb457add9f7829fbaf40680d70814ee4
Closes-Bug: #1231850