Clean up cookbook for bp/clean-up-attr-for-mq-and-db
This change removes the attributes we'll be moving to openstack-common
and changes reference of those attributes to reflect new (more
consistent) namespace. We also fix a test where node isn't being
converged after setting attributes and import chefspec/berksfile in
spec/spec_helper.rb so we can run rspec outside of strainer.

Change-Id: I7e27a60c4245de2e4d8033439a2429d89006a356
Implements: blueprint clean-up-attr-for-mq-and-db