Wait for dnsmasq to become available
Fixes: bug 1280214

When we install dnsmasq, it takes a few seconds to start up, during
which time dns resolution on the system is unavailable.  If we carry on
with the chef run during this time, subsequent steps (eg package
installations) will fail due to transient errors with name resolution.

This patch introduces a small loop to ensure we have working dns resolution
before moving on

Change-Id: Ia508a7243554bca64c0451d99086ae596ec9e462