Point to correct lbaas_agent.ini file
On CentOS / Ubuntu, lbaas agent fails to start as we drop in a config
in a non-standard location.  The stock distro-provided config is also
problematic as it doesn't have a default interface_driver set which
prevents the service from starting.

This change defaults all distros to use /etc/neutron/lbaas_agent.ini
and tidies up ownership/mode of the dropped file.

Change-Id: I919e535989c5ec05ee196235fbd528eb971cbb05
Closes-Bug: #1283237