Subscribe Neutron agent services to neutron.conf
Before this change, a change to neutron.conf would not cause any agent
to restart. I doubt this was intended, since for example changing
rabbitmq stuff in neutron.conf really needs the agent to restart to have
any effect.

Furthermore, the balancer recipe subscription/notification logic has
been changed to fit in line with the other agents, i.e. having the ini
template notify the service and the service subscribe to neutron.conf.

Change-Id: I2d5b20a13945fb8a360b81091e68e0fcf5748e93
Closes-Bug: #1300076