Increase IP binding flexibility changes
Changes have been made to the common cookbook so that if a
bind_interface is specified for an endpoint, then the IP is looked up
for that endpoint using address_for, and if no bind_interface is
specified then the value of the host attribute of the endpoint is used
instead.

This change set removes code around the network API endpoint that
implements this exact behavior, which is no longer needed since it is
now provided by the common cookbook.  This change set also replaces
direct calls to address_for by the OVS and LinuxBridge code with
endpoints.  This allows the user to either specify an interface or an
IP to bind to using the standard endpoint library for these two
plugins.

This change is dependent on change: https://review.openstack.org/83859

blueprint increase-ip-binding-flexibility

Change-Id: Ie0aa07ab511c4ccbe9686da4dfb05934f9708b9a