Allow binding locally to different IP than endpoint IP
Sometimes you want the actual endpoint IP to be a load balanced IP, but
the bind IP to be a local IP (different per node).

This change allows that by using attributes from the [network-api-bind]
namespace, in addition to standardising on IP discovery by using the 'endpoint'
method as used in the other cookbooks.

Implements blueprint increase-ip-binding-flexibility

Change-Id: I2eecb7601a426e3611d5c9c4a10b49863fc4e48d