Allow policy.json file to be specified
The policy.json file was removed in 853fa9fcc6d7e9e6607be188299bf6c9208f24f7
this will allow you to specify an attribute to point to a location of a
policy.json file to be used.

By default this is nil and the packaged policy.json file will be used,
but if required this can be overridden.

Change-Id: I2d21211e24e64d7046cf3782400310eb1bff4500
Implements: blueprint remote-policy