Fix openvswitch and linux bridge agent confs for ml2 on redhat
The agent conf files were not being created when using ml2 on redhat.
This patch creates the conf directory and file for the agent recipes.
For openvsiwtch, it's specific to redhat, as debian uses different conf
file, reported under this bug 1314751.
For linuxbridge, it's the same conf for both redhat and ubuntu.

See bug for more details.

Change-Id: Iae8fff68c5b270593ba9c23ecc8deac7577e2382
Closes-Bug: #1311311