Fix symlink when main changed from ML2 to ovs
When the main plugin is changed from ML2 to OpenVSwitch, the redhat
required plugins.ini symlink does not get properly updated. Change the
symlink to rely solely on the main plugin attribute and not the conf
files as there is not a 1-1 relationship between them.  See bug for more
details.

Change-Id: If32747e54e8a8f31d1d553b750e6ae663670b873
Closes-Bug: #1320011