Add support for removing all l3 routers from agent
The neutron-ha-tool.py tool has the ability to rebalance routers on l3
agents, this adds the ability to remove all the routers from one l3
agent so it can be taken down for maintenance.  Just something that is a
minor change but quite handy to have available.

Also added the ability to read the region, if your neutron needs it.

Change-Id: I17d8f6f8c707dd6be0e64ccc2d5637b72b8ed392