Fix ubuntu ml2_conf missing ovs section
Used the existing ovs conf as a partial template to simply include
in the ml2 conf.  This keeps all the attributes the same.
A more detailed spec will have to be done after some refactoring
to allow for sections to be tested.

Please give this idea a try, it seems to allow me to setup neutron
on ubuntu with ml2. I'm working toward using this as a basis for
getting the repo aio_neutron test working.

Change-Id: Ia988f3d7ef1c280c40b55cea4f455710f673dd32
Closes-Bug: #1314751