Resolve the db migration dependency in Neutron
Most of the code of advance services is splited from Neutron,
except create db tables. So there will be errors when execute db
migration of advance services before Neutron. It is a dependency
of db migration.
To avoid erros caused by db dependency, create a recipe to handle
all the db migration for Neutron and advance services. With that
the advance services db migration can be handle after related
tables are created by Neutron db migration.

Change-Id: I502976f34977fae4f2df831382593560a56caa92
Closes-Bug: #1438148