Make the metadata_workers attribute in metadata_agent.ini configurable.
The default value of metadata_workers in metadata_agent.ini is half of the
number of CPU cores in the computer system. The value will be large only if
there are many CPU cores, which may not be necessary. It should be acceptable
to make this attribute configurable, which follows what has been done for
api_workers and rpc_workers in neutron.conf.

Change-Id: I823b485d72fb74c13e4bce221a256cfed6770d65