Restart neutron-vpn-agent in l3_agent recipe when vpn is enabled
As there are some potential problems when neutron-l3-agent and
neutron-vpn-agent services are both running in the network node,
we should disable and stop neutron-l3-agent service.

The l3_agent recipe will be executed when running vpn_agent recipe,
so when firewall is enabled, restart neutron-vpn-agent service
instead of neutron-l3-agent service.

Closes-Bug: #1452121
Change-Id: I3d9ae89994770846094e24b3355ffad1a4e0d9ef