Install lbaas agent pkg only when needed
The lbaas recipe is meant to be run on a controller node, while the
lbaas_agent recipe is for a network node. Installing the lbaas agent
package on a controller leads to a failing service being started.
In order to avoid that, we need to split up the package lists and
install the agent only when it is needed.
Installing the lbaas agent on a controller node instead of a network

Change-Id: I8742e59bf5a07f1e26461e29f3cb2a1267b20925